Separate Node and NodeMeta
